[test] move two tests from from stdlib to runtime
